Release 0..9 Alpha
Changed the parser to use a TextElementEnumerator and a Tokenizer to
read the file. More support for Globalization (supports full Unicode
chars in names and culture-specific numbers). Changed LuaTable length to
get the highest number. Fixed inline comments. Moved the Parser.Items
namespace to the Parser namespace to remove the empty namespace for the
developer.

Added support for base type access in derived types, for params arrays
and optional arguments for C# functions, for converting a LuaTable to an
array in method arguments. Added a changes and license file.