Rewrote the runtime, again.
This is the last commit from the dark ages.  Mostly completed long
ago and simply cleaned up and pushed.  I will be better about making
smaller commits with better descriptions as well as making changes
to make the code easier to read and maintain.

This change will probably be the last large one that breaks
everything.  Future commits will be more self-contained and will
focus on one area.  Because of the lack of tests, I am unsure of
what this will break.

This removes support for (will be fixed later):
* Controlling accessibility through LuaUserData.
* Operator handling of C# types within Lua.
* Proper tail calls.

The new runtime is based on having wrapped values rather than
passing raw objects around.  Now all values within Lua are objects
that implement ILuaValue and define methods such as Arithmetic and
GetIndex.  This moves most of the runtime into the objects themselves
and allows inheritance to handle much of the checking of types.

This also move most of the runtime code into ModMaker.Lua which can
be shared with the portable version.  I am considering removing the
portable library support.

Closes #2
Closes #11