Change license to Apache License.
This changes from the restrictive GPL license to the more permissive
Apache License.  This also adds a license header to the source files.
The years were pulled from the git history, with some adjustments.
In some cases, the git history was incorrect because this repo has
moved around and changed a lot, so git may not have detected the
the file moves.