Remove support for specifying overload
This removes the extra language feature to explicitly specify which
overload to call.  This was an odd feature that required a lot of extra
work to plumb the value through.  By removing it, this simplifies the
internal calling convention.

To replace it, the Lua function "overload" now can select which overload
to call.  For example:

    overload(myFunc, 2)(args...)