Remove several methods from ILuaRuntime
As part of removing the runtime, this removes many of the Create*
methods and uses the direct versions instead.  The runtime was intended
to be replaced by the app, but this probably won't happen.  The app
probably wants to use our LuaValue types and won't want to implement
thier own.  Also, even if they did, much of our code assumes using our
types.  So this removes the Create* methods and just calls our versions
directly.  This will also increase performance by avoiding several
virtual calls.

This also fixes several bugs in the generated code.