Add multi-target for .Net framework 4.5/4.8.
Now we also target .Net framework versions 4.5 and 4.8 in addition to
.Net Core 3.1.  This allows people using older versions to use this
library and allow us to embed debug info in the generated code.  We
can't do it in .Net core since it isn't exposed there.

We still target C# 8.0, so we still have the same language features,
with the exception of some things like "^1" and "25..." indexing.  But
we still can use nullable reference types.