Split CompilerMessage and exception
This adds a separate exception type from the CompilerMessage.  This
allows construction of an exception based on multiple messages.  This
change also adds a collection type to help with holding onto messages
and throwing the exception.

For the most part, the parser still only returns one error since we
can't recover; but this sets the ground work for warnings and for
multiple errors in a file.