Remove "self" from calling convention
Before, the "self" object would be passed as a separate argument within
the internal calling convention.  This would allow some cases of fixing
incorrect usages.  However, this complicates the calling convention for
a small feature that doesn't really follow the Lua spec.  To simplify
things, now the "self" argument is passed in the main arguments like
the spec specifies.